global error handler

+++ src/main/java/com/example/webapp/service/MatchService.java
@@ -2,6 +2,7 @@ package com.example.webapp.service;
 
 import com.example.webapp.api.Match;
 import com.example.webapp.exception.DateInPastException;
+import com.example.webapp.exception.MatchNotFoundException;
 import com.example.webapp.repository.MatchEntity;
 import com.example.webapp.repository.MatchRepository;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +18,9 @@ public class MatchService {
 
     private final MatchRepository repository;
 
+    public boolean checkIfMatchExists (Long id){
+        return repository.existsById(id);
+    }
     public void create(Match match) {
         if (match.getFirstTeam().isEmpty() || match.getSecondTeam().isEmpty()) {
             throw new IllegalStateException("Nie podano zespołów biorących udział w meczu.");
@@ -24,6 +28,9 @@ public class MatchService {
 
         if (LocalDateTime.now().isAfter(match.getStartTime())){
             throw new DateInPastException("Godzina meczu jest z przeszłości.");
+        }
+        if (!repository.existsById(match.getId())){
+            throw new MatchNotFoundException("Mecz nie istnieje.");
         }
 
         repository.save(MatchEntity.builder()
@@ -48,6 +55,9 @@ public class MatchService {
     }
 
     public void delete(Long id) {
+        if (!repository.existsById(id)){
+            throw new MatchNotFoundException("Mecz nie istnieje.");
+        }
         repository.delete(id);
 
     }
